Make ucodegen
