+++ src/ucodegen/UcodeCodeGen.java
@@ -1,4 +1,6 @@
+package ucodegen;
 
+import antlrMiniC.*;
 import org.antlr.v4.runtime.ANTLRFileStream;
 import org.antlr.v4.runtime.CommonTokenStream;
 import org.antlr.v4.runtime.tree.ParseTree;
